refactor(stats-writer): split stats_writer.rs by responsibility (#1129)
Behavior-preserving module split of the 2,564-line stats_writer.rs into
stats_writer/{mod,metadata,marshal,components,serialization_header}.rs.

- serialization_header.rs: the SERIALIZATION_HEADER builder (the headline
  extraction) + its 5 tests.
- components.rs: VALIDATION/COMPACTION/STATS (nb + da) builders,
  EstimatedHistogram + TombstoneHistogram serialisers + 5 tests.
- metadata.rs: StatisticsMetadata aggregation + TombstoneHistogram + 12 tests.
- marshal.rs: cql_type_to_marshal_type + helpers + 1 test.
- mod.rs: StatisticsWriter orchestration (TOC + write), constants, re-exports
  (StatisticsMetadata, StatisticsWriter, TombstoneHistogram, and pub(crate)
  cql_type_to_marshal_type so data_writer.rs's path is unchanged) + 5 tests.

Public surface and on-disk bytes unchanged. Tests relocated 1:1 (28 total,
count preserved). No semantic changes.

Epic #1116.
